+++ src/_internals/createPath.js
@@ -1,7 +1,3 @@
-import { isInteger } from "../../source/_internals/isInteger"
-
 export function createPath(path, delimiter = '.'){
-  return typeof path === 'string' ?
-  path.split(delimiter).map(x => isInteger(Number(x)) ? Number(x) : x) :
-  path
+  return typeof path === 'string' ? path.split(delimiter) : path.map(String)
 }